defi-audit/requirements.txt
fastapi
uvicorn
web3
python-dotenv